Fix the bug that ov.utils.plot_paga layout parameter doesn't work when set basis.
The layout parameter of ov.utils.plot_paga does not work when you set the basis parameter, like `ov.utils.plot_paga(adata, basis='umap', layout="rt")`. This should be an issue with scvelo, but to make ov run correctly, an if statement is added here to ensure it works properly.